[Design] #35 - 헤더 뷰 추가
