+++ ShortWeather/ShortWeather/Presentation/WeekWeather/View/WeekWeatherUIView.swift
@@ -4,6 +4,7 @@
 //
 //  Created by KJ on 2023/01/09.
 //
+
 import UIKit
 
 import Moya
@@ -13,11 +14,7 @@ import Then
 final class WeekWeatherUIView: UIView {
     
     // MARK: - UI Components
-
-    private let weekDescriptionView: UIView = UIView()
-    private let dailyForecastLabel: UILabel = UILabel()
-    private let timeZoneLabel: UILabel = UILabel()
-    private let temperatureLabel: UILabel = UILabel()
+    
     private let weekWeatherCollectionView: UICollectionView = {
         let layout = UICollectionViewFlowLayout()
         layout.scrollDirection = .vertical
@@ -34,7 +31,6 @@ final class WeekWeatherUIView: UIView {
     
     override init(frame: CGRect) {
         super.init(frame: frame)
-        print("ddd")
         setUI()
         setLayout()
         setDelegate()
@@ -50,67 +46,24 @@ extension WeekWeatherUIView {
     // MARK: - UI Components Property
     
     private func setUI() {
-        dailyForecastLabel.do {
-            $0.text = "일별예보"
-            $0.font = .fontGuide(.caption1)
-            $0.textColor = Color.gray7
-        }
-        
-        timeZoneLabel.do {
-            $0.text = "오전 / 오후"
-            $0.font = .fontGuide(.caption1)
-            $0.textColor = Color.gray7
-        }
-        
-        temperatureLabel.do {
-            $0.text = "최저 / 최고"
-            $0.font = .fontGuide(.caption1)
-            $0.textColor = Color.gray7
-        }
-        
-        weekDescriptionView.do {
-            $0.backgroundColor = .clear
-        }
-        
         weekWeatherCollectionView.do {
             $0.registerCells(WeekWeatherCollectionViewCell.self)
             $0.backgroundColor = .clear
+            $0.register(WeekWeatherHeaderView.self, forSupplementaryViewOfKind: UICollectionView.elementKindSectionHeader, withReuseIdentifier: "WeekWeatherHeaderView")
         }
     }
     
     // MARK: - Layout Helper
     
     private func setLayout() {
-        addSubviews(weekDescriptionView, weekWeatherCollectionView)
-        weekDescriptionView.addSubviews(dailyForecastLabel, timeZoneLabel, temperatureLabel)
-        
-        weekDescriptionView.snp.makeConstraints {
-            $0.top.equalToSuperview().offset(16)
-            $0.leading.equalToSuperview().offset(36)
-            $0.centerX.equalToSuperview()
-            $0.height.equalTo(18)
-        }
-        
-        dailyForecastLabel.snp.makeConstraints {
-            $0.centerY.equalToSuperview()
-            $0.leading.equalToSuperview()
-        }
-        
-        timeZoneLabel.snp.makeConstraints {
-            $0.centerY.equalToSuperview()
-            $0.centerX.equalToSuperview()
-        }
-        
-        temperatureLabel.snp.makeConstraints {
-            $0.trailing.equalToSuperview()
-            $0.centerY.equalToSuperview()
-        }
+        addSubviews(weekWeatherCollectionView)
         
         weekWeatherCollectionView.snp.makeConstraints {
-            $0.top.equalTo(weekDescriptionView.snp.bottom).offset(4)
+            // 수정해야함
+            $0.top.equalToSuperview().offset(4)
             $0.leading.equalToSuperview().offset(28)
             $0.trailing.equalToSuperview().inset(28)
-            $0.bottom.equalToSuperview().offset(20)
+            $0.bottom.equalToSuperview().inset(20)
         }
     }
     
@@ -130,6 +83,11 @@ extension WeekWeatherUIView: UICollectionViewDelegateFlowLayout {
         let width = bounds.size.width - 56
         return CGSize(width: width, height: 72)
     }
+    
+    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, referenceSizeForHeaderInSection section: Int) -> CGSize {
+        let width = collectionView.frame.width
+        return CGSize(width: width, height: 18)
+    }
 }
 
 extension WeekWeatherUIView: UICollectionViewDataSource {
@@ -141,5 +99,15 @@ extension WeekWeatherUIView: UICollectionViewDataSource {
     func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
         let cell = collectionView.dequeueCell(type: WeekWeatherCollectionViewCell.self, indexPath: indexPath)
         return cell
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, viewForSupplementaryElementOfKind kind: String, at indexPath: IndexPath) -> UICollectionReusableView {
+        guard kind == UICollectionView.elementKindSectionHeader,
+              let header = collectionView.dequeueReusableSupplementaryView(
+                ofKind: kind,
+                withReuseIdentifier: "WeekWeatherHeaderView",
+                for: indexPath
+              ) as? WeekWeatherHeaderView else { return UICollectionReusableView() }
+        return header
     }
 }
